requirements.txt

fastapi==0.115.2
uvicorn==0.32.0
streamlit==1.39.0
langchain==0.3.3
langchain-community==0.3.2
langchain-groq==0.2.0
sentence-transformers==3.2.1
chromadb==0.5.13
pyjwt==2.9.0
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.3
redis==5.1.1
passlib[bcrypt]==1.7.4

